+++ mipac/manager/mute.py
@@ -1,0 +1,23 @@
+from mipac.abstract.manager import AbstractManager
+from mipac.actions.mute import MuteActions
+from mipac.http import HTTPClient
+from mipac.manager.client import ClientActions
+
+
+class MuteManager(AbstractManager):
+    def __init__(
+        self,
+        user_id: str | None = None,
+        *,
+        session: HTTPClient,
+        client: ClientActions
+    ) -> None:
+        self._user_id: str | None = user_id
+        self._session: HTTPClient = session
+        self._client: ClientActions = client
+
+    @property
+    def action(self) -> MuteActions:
+        return MuteActions(
+            user_id=self._user_id, session=self._session, client=self._client
+        )